+++ src/domain/community/virtualContributor/knowledgeBase/VCKnowledgeBaseRoute.tsx
@@ -3,7 +3,7 @@ import { Route, Routes } from 'react-router-dom';
 import VCProfilePage from '../vcProfilePage/VCProfilePage';
 import CalloutPage from '@/domain/collaboration/CalloutPage/CalloutPage';
 import { buildVCKnowledgeBaseUrl } from '@/main/routing/urlBuilders';
-import { virtualContributorsCalloutRestrictions } from '../utils/virtualContributorsCalloutRestrictions';
+import { virtualContributorsCalloutRestrictions } from './virtualContributorsCalloutRestrictions';
 
 export const VCKnowledgeBaseRoute = () => {
   return (
